Adapting services to hibernate session changes DSA-108

+++ src/main/java/sepm/dsa/service/TraderCategoryServiceImpl.java
@@ -6,6 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.transaction.annotation.Transactional;
 import sepm.dsa.dao.TraderCategoryDao;
 import sepm.dsa.exceptions.DSAValidationException;
+import sepm.dsa.model.Trader;
 import sepm.dsa.model.TraderCategory;
 
 import javax.validation.ConstraintViolation;
@@ -14,13 +15,15 @@ import javax.validation.Validator;
 import java.util.List;
 import java.util.Set;
 
+@Transactional(readOnly = true)
 public class TraderCategoryServiceImpl implements TraderCategoryService {
     private static final Logger log = LoggerFactory.getLogger(TraderCategoryServiceImpl.class);
     private Validator validator = Validation.byProvider(HibernateValidator.class).configure().buildValidatorFactory().getValidator();
 
     private TraderCategoryDao traderCategoryDao;
+	private TraderService traderService;
 
-    @Override
+	@Override
     public TraderCategory get(int id) {
         log.debug("calling get(" + id + ")");
         TraderCategory result = traderCategoryDao.get(id);
@@ -47,8 +50,14 @@ public class TraderCategoryServiceImpl implements TraderCategoryService {
     @Override
     @Transactional(readOnly = false)
     public void remove(TraderCategory t) {
-        log.debug("calling removeConnection(" + t + ")");
-        traderCategoryDao.remove(t);
+        log.debug("calling remove(" + t + ")");
+
+	    List<Trader> traders = traderService.getAllByCategory(t);
+	    if (traders.isEmpty()) {
+		    traderCategoryDao.remove(t);
+	    } else {
+		    throw new DSAValidationException("Löschen nicht möglich. Zu dieser Kategorie sind noch Händler vorhanden.");
+	    }
     }
 
     @Override
@@ -75,4 +84,8 @@ public class TraderCategoryServiceImpl implements TraderCategoryService {
             throw new DSAValidationException("Händlerkategorie ist nicht valide.", violations);
         }
     }
+
+	public void setTraderService(TraderService traderService) {
+		this.traderService = traderService;
+	}
 }
